requirements.txt
sentence-transformers
numpy
langchain-openai
langchain-core
crawl4ai
python-dotenv
